clamp stuff

+++ src/assembler/ArmV8Assembler.hpp
@@ -104,11 +104,11 @@ namespace tulip::hook {
 		void blr(ArmV8Register reg);
 		void nop();
 
-		void ldr(ArmV8Register dst, ArmV8Register src, int16_t imm);
+		void ldr(ArmV8Register dst, ArmV8Register src, int32_t imm);
 		void ldr(ArmV8Register dst, int32_t literal);
 
-		void tbz(ArmV8Register reg, uint8_t bit, int16_t imm);
-		void tbnz(ArmV8Register reg, uint8_t bit, int16_t imm);
+		void tbz(ArmV8Register reg, int32_t bit, int32_t imm);
+		void tbnz(ArmV8Register reg, int32_t bit, int32_t imm);
 
 		/* Pseudo instructions */
